requirements.txt
flask==2.0.1
pillow==8.3.1
numpy==1.21.0
tensorflow==2.5.0
tflite-runtime==2.5.0
werkzeug==2.0.1 